requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic_settings
authlib
python-jose[cryptography]
passlib